# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InterviewAnswer {
  questionId: string;
  answer: string;
}

export type DatabaseConfigEngine = typeof DatabaseConfigEngine[keyof typeof DatabaseConfigEngine];


export const DatabaseConfigEngine = {
  sqlite: 'sqlite',
  postgresql: 'postgresql',
  supabase: 'supabase',
  firebase: 'firebase',
  none: 'none',
} as const;

export type DatabaseConfigPersistence = typeof DatabaseConfigPersistence[keyof typeof DatabaseConfigPersistence];


export const DatabaseConfigPersistence = {
  local: 'local',
  server: 'server',
  hybrid: 'hybrid',
} as const;

export interface DatabaseConfig {
  engine: DatabaseConfigEngine;
  persistence: DatabaseConfigPersistence;
  schemaNotes: string;
  authRequired: boolean;
}

export interface InterviewRequest {
  prompt: string;
  answers: InterviewAnswer[];
  database: DatabaseConfig;
}

export type InterviewResponseKind = typeof InterviewResponseKind[keyof typeof InterviewResponseKind];


export const InterviewResponseKind = {
  text: 'text',
  choice: 'choice',
} as const;

export interface InterviewResponse {
  done: boolean;
  questionId?: string;
  question?: string;
  kind?: InterviewResponseKind;
  options?: string[];
  progress: number;
}

export type BlueprintFeaturePriority = typeof BlueprintFeaturePriority[keyof typeof BlueprintFeaturePriority];


export const BlueprintFeaturePriority = {
  must: 'must',
  should: 'should',
  later: 'later',
} as const;

export interface BlueprintFeature {
  name: string;
  description: string;
  priority: BlueprintFeaturePriority;
}

export interface BlueprintRequest {
  prompt: string;
  answers: InterviewAnswer[];
  database: DatabaseConfig;
}

export interface BlueprintResponse {
  title: string;
  summary: string;
  platform: string;
  database: DatabaseConfig;
  features: BlueprintFeature[];
  openQuestions: string[];
  generatedAt: string;
}

export interface GithubRepository {
  id: number;
  fullName: string;
  defaultBranch: string;
  private: boolean;
  htmlUrl: string;
}

export interface GithubRepositoriesResponse {
  connected: boolean;
  repositories: GithubRepository[];
}

export interface GithubWorkflow {
  id: number;
  name: string;
  path: string;
  state: string;
}

export interface GithubWorkflowsResponse {
  workflows: GithubWorkflow[];
}

export interface GithubPushRequest {
  path: string;
  content: string;
  message: string;
  branch: string;
  sha?: string;
}

export type GithubDispatchRequestInputs = {[key: string]: string};

export interface GithubDispatchRequest {
  workflowId: string;
  ref: string;
  inputs?: GithubDispatchRequestInputs;
}

export interface GithubWriteResponse {
  ok: boolean;
  message: string;
  url?: string;
}

export interface GithubRun {
  id: number;
  name: string;
  status: string;
  conclusion?: string;
  htmlUrl: string;
}

export interface GithubRunsResponse {
  runs: GithubRun[];
}