# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: forge
    description: Local-first app building workflow
  - name: github
    description: Authenticated GitHub repository and Actions controls
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /forge/interview:
    post:
      operationId: createInterviewQuestion
      tags: [forge]
      summary: Create the next app-building interview question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewRequest"
      responses:
        "200":
          description: Next question and completion state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewResponse"
        "400":
          description: Invalid interview input
  /forge/blueprint:
    post:
      operationId: createBlueprint
      tags: [forge]
      summary: Validate and normalize an app build brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlueprintRequest"
      responses:
        "200":
          description: Normalized blueprint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlueprintResponse"
        "400":
          description: Invalid blueprint input
  /forge/github/repos:
    get:
      operationId: listGithubRepositories
      tags: [github]
      summary: List repositories available to the connected GitHub account
      responses:
        "200":
          description: Accessible repositories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubRepositoriesResponse"
  /forge/github/repos/{owner}/{repo}/workflows:
    get:
      operationId: listGithubWorkflows
      tags: [github]
      summary: List Actions workflows in a repository
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository workflows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubWorkflowsResponse"
  /forge/github/repos/{owner}/{repo}/push:
    post:
      operationId: pushGithubFile
      tags: [github]
      summary: Push one generated project file to GitHub
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPushRequest"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubWriteResponse"
  /forge/github/repos/{owner}/{repo}/dispatch:
    post:
      operationId: dispatchGithubWorkflow
      tags: [github]
      summary: Start a GitHub Actions workflow
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubDispatchRequest"
      responses:
        "202":
          description: Workflow accepted by GitHub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubWriteResponse"
  /forge/github/repos/{owner}/{repo}/runs:
    get:
      operationId: listGithubRuns
      tags: [github]
      summary: List recent GitHub Actions runs
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent workflow runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubRunsResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    InterviewAnswer:
      type: object
      properties:
        questionId:
          type: string
        answer:
          type: string
      required:
        - questionId
        - answer
    InterviewRequest:
      type: object
      properties:
        prompt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/InterviewAnswer"
        database:
          $ref: "#/components/schemas/DatabaseConfig"
      required:
        - prompt
        - answers
        - database
    InterviewResponse:
      type: object
      properties:
        done:
          type: boolean
        questionId:
          type: string
        question:
          type: string
        kind:
          type: string
          enum: [text, choice]
        options:
          type: array
          items:
            type: string
        progress:
          type: integer
      required:
        - done
        - progress
    BlueprintFeature:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [must, should, later]
      required:
        - name
        - description
        - priority
    BlueprintRequest:
      type: object
      properties:
        prompt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/InterviewAnswer"
        database:
          $ref: "#/components/schemas/DatabaseConfig"
      required:
        - prompt
        - answers
        - database
    BlueprintResponse:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        platform:
          type: string
        database:
          $ref: "#/components/schemas/DatabaseConfig"
        features:
          type: array
          items:
            $ref: "#/components/schemas/BlueprintFeature"
        openQuestions:
          type: array
          items:
            type: string
        generatedAt:
          type: string
      required:
        - title
        - summary
        - platform
        - database
        - features
        - openQuestions
        - generatedAt
    DatabaseConfig:
      type: object
      properties:
        engine:
          type: string
          enum: [sqlite, postgresql, supabase, firebase, none]
        persistence:
          type: string
          enum: [local, server, hybrid]
        schemaNotes:
          type: string
        authRequired:
          type: boolean
      required:
        - engine
        - persistence
        - schemaNotes
        - authRequired
    GithubRepository:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        defaultBranch:
          type: string
        private:
          type: boolean
        htmlUrl:
          type: string
      required:
        - id
        - fullName
        - defaultBranch
        - private
        - htmlUrl
    GithubRepositoriesResponse:
      type: object
      properties:
        connected:
          type: boolean
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/GithubRepository"
      required:
        - connected
        - repositories
    GithubWorkflow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
        state:
          type: string
      required:
        - id
        - name
        - path
        - state
    GithubWorkflowsResponse:
      type: object
      properties:
        workflows:
          type: array
          items:
            $ref: "#/components/schemas/GithubWorkflow"
      required:
        - workflows
    GithubPushRequest:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        message:
          type: string
        branch:
          type: string
        sha:
          type: string
      required:
        - path
        - content
        - message
        - branch
    GithubDispatchRequest:
      type: object
      properties:
        workflowId:
          type: string
        ref:
          type: string
        inputs:
          type: object
          additionalProperties:
            type: string
      required:
        - workflowId
        - ref
    GithubWriteResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        url:
          type: string
      required:
        - ok
        - message
    GithubRun:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        conclusion:
          type: string
        htmlUrl:
          type: string
      required:
        - id
        - name
        - status
        - htmlUrl
    GithubRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: "#/components/schemas/GithubRun"
      required:
        - runs