# lib/api-zod/src/generated/types/blueprintRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DatabaseConfig } from './databaseConfig';
import type { InterviewAnswer } from './interviewAnswer';

export interface BlueprintRequest {
  prompt: string;
  answers: InterviewAnswer[];
  database: DatabaseConfig;
}