# lib/api-zod/src/generated/types/databaseConfig.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DatabaseConfigEngine } from './databaseConfigEngine';
import type { DatabaseConfigPersistence } from './databaseConfigPersistence';

export interface DatabaseConfig {
  engine: DatabaseConfigEngine;
  persistence: DatabaseConfigPersistence;
  schemaNotes: string;
  authRequired: boolean;
}