# lib/api-zod/src/generated/types/githubDispatchRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubDispatchRequestInputs } from './githubDispatchRequestInputs';

export interface GithubDispatchRequest {
  workflowId: string;
  ref: string;
  inputs?: GithubDispatchRequestInputs;
}