# lib/api-zod/src/generated/types/githubPushRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubPushRequest {
  path: string;
  content: string;
  message: string;
  branch: string;
  sha?: string;
}