# lib/api-zod/src/generated/types/githubRepositoriesResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubRepository } from './githubRepository';

export interface GithubRepositoriesResponse {
  connected: boolean;
  repositories: GithubRepository[];
}