# lib/api-zod/src/generated/types/githubRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubRun {
  id: number;
  name: string;
  status: string;
  conclusion?: string;
  htmlUrl: string;
}