# lib/api-zod/src/generated/types/githubRunsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubRun } from './githubRun';

export interface GithubRunsResponse {
  runs: GithubRun[];
}