# lib/api-zod/src/generated/types/githubWorkflow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubWorkflow {
  id: number;
  name: string;
  path: string;
  state: string;
}