# lib/api-zod/src/generated/types/githubWorkflowsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubWorkflow } from './githubWorkflow';

export interface GithubWorkflowsResponse {
  workflows: GithubWorkflow[];
}