# lib/api-zod/src/generated/types/githubWriteResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubWriteResponse {
  ok: boolean;
  message: string;
  url?: string;
}